# backend/internal/api/api.go
package api

import (
	"context"
	"net/http"
	"os"
	"path"
	"path/filepath"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"s3desk/internal/bucketgov"
	"s3desk/internal/config"
	"s3desk/internal/jobs"
	"s3desk/internal/metrics"
	"s3desk/internal/store"
	"s3desk/internal/ws"
)

type Dependencies struct {
	Config          config.Config
	Store           *store.Store
	Jobs            *jobs.Manager
	Hub             *ws.Hub
	Metrics         *metrics.Metrics
	ServerAddr      string
	ShutdownContext context.Context
}

func New(dep Dependencies) http.Handler {
	r := chi.NewRouter()
	r.Use(middleware.RequestID)
	r.Use(requestBodyIdleTimeout)

	api := &server{
		cfg:             dep.Config,
		store:           dep.Store,
		jobs:            dep.Jobs,
		hub:             dep.Hub,
		metrics:         dep.Metrics,
		serverAddr:      dep.ServerAddr,
		shutdownContext: dep.ShutdownContext,
		proxySecret:     resolveProxySecret(dep.Config.APIToken),
		realtimeTickets: newRealtimeTicketStore(30 * time.Second),
		authLimit:       newAuthFailureLimiter(10, time.Minute, time.Minute),
		uploadLimit:     newRequestLimiter(dep.Config.UploadMaxConcurrentRequests),
		realtimeLimit:   newRequestLimiter(defaultRealtimeMaxConnections),
		realtimeMax:     defaultRealtimeMaxConnections,
		bucketGov: bucketgov.NewServiceWithOptions(
			bucketgov.NewDefaultRegistryWithOptions(bucketgov.DefaultRegistryOptions{AllowRemote: dep.Config.AllowRemote}),
			bucketgov.ServiceOptions{AllowRemote: dep.Config.AllowRemote},
		),
	}

	r.Use(api.allowOnlySafeMethods)
	r.Use(api.requestLogger)
	r.Use(middleware.Recoverer)
	r.Use(middleware.Compress(5))
	r.Use(securityHeaders)

	apiRouter := chi.NewRouter()
	apiRouter.Use(api.requireLocalHost)
	apiRouter.Use(api.cors)
	apiRouter.Use(api.requireAPIToken)

	apiRouter.Get("/ws", api.handleWSUpgrade)
	apiRouter.Get("/events", api.handleEventsSSE)
	apiRouter.Post("/realtime-ticket", api.handleCreateRealtimeTicket)
	apiRouter.Get("/bootstrap", api.handleGetBootstrap)
	apiRouter.Get("/meta", api.handleGetMeta)
	apiRouter.Get("/server/backup", api.handleGetServerBackup)
	apiRouter.Post("/server/backup/transfer", api.handleTransferServerBackup)
	apiRouter.Post("/server/restore", api.handleRestoreServerBackup)
	apiRouter.Post("/server/restore/transfer", api.handleTransferServerRestore)
	apiRouter.Post("/server/import-portable/preview", api.handlePreviewPortableImport)
	apiRouter.Post("/server/import-portable", api.handleImportPortableBackup)
	apiRouter.Get("/server/restores", api.handleListServerRestores)
	apiRouter.Delete("/server/restores/{restoreId}", api.handleDeleteServerRestore)
	apiRouter.Get("/jobs/{jobId}/logs", api.handleGetJobLogs)
	apiRouter.With(api.requireStoredProfile).Get("/jobs", api.handleListJobs)
	apiRouter.With(api.requireStoredProfile).Get("/jobs/{jobId}", api.handleGetJob)
	apiRouter.With(api.requireStoredProfile).Get("/jobs/{jobId}/artifact", api.handleGetJobArtifact)
	apiRouter.With(api.requireStoredProfile).Get("/buckets/{bucket}/objects/search", api.handleSearchObjects)
	apiRouter.With(api.requireStoredProfile).Get("/buckets/{bucket}/objects/index-summary", api.handleGetObjectIndexSummary)
	apiRouter.With(api.requireStoredProfile).Post("/buckets/{bucket}/objects/favorites", api.handleCreateObjectFavorite)
	apiRouter.With(api.requireStoredProfile).Delete("/buckets/{bucket}/objects/favorites", api.handleDeleteObjectFavorite)
	apiRouter.With(api.requireStoredProfile).Get("/local/entries", api.handleListLocalEntries)
	apiRouter.Post("/uploads/{uploadId}/files", api.handleUploadFiles)

	apiRouter.Route("/profiles", func(r chi.Router) {
		r.Get("/", api.handleListProfiles)
		r.Post("/", api.handleCreateProfile)
		r.Route("/{profileId}", func(r chi.Router) {
			r.Patch("/", api.handleUpdateProfile)
			r.Delete("/", api.handleDeleteProfile)
			r.Post("/test", api.handleTestProfile)
			r.Post("/benchmark", api.handleBenchmarkProfile)
			r.Get("/export", api.handleExportProfile)
			r.Get("/tls", api.handleGetProfileTLS)
			r.Put("/tls", api.handlePutProfileTLS)
			r.Delete("/tls", api.handleDeleteProfileTLS)
		})
	})

	apiRouter.Group(func(r chi.Router) {
		r.Use(api.requireProfile)

		r.Route("/buckets", func(r chi.Router) {
			r.Get("/", api.handleListBuckets)
			r.Post("/", api.handleCreateBucket)
			r.Delete("/{bucket}", api.handleDeleteBucket)
		})

		r.Route("/buckets/{bucket}/policy", func(r chi.Router) {
			r.Get("/", api.handleGetBucketPolicy)
			r.Put("/", api.handlePutBucketPolicy)
			r.Delete("/", api.handleDeleteBucketPolicy)
			r.Post("/validate", api.handleValidateBucketPolicy)
		})

		r.Route("/buckets/{bucket}/governance", func(r chi.Router) {
			r.Get("/", api.handleGetBucketGovernance)
		})

		r.Route("/buckets/{bucket}/governance/access", func(r chi.Router) {
			r.Get("/", api.handleGetBucketAccess)
			r.Put("/", api.handlePutBucketAccess)
		})

		r.Route("/buckets/{bucket}/governance/public-exposure", func(r chi.Router) {
			r.Get("/", api.handleGetBucketPublicExposure)
			r.Put("/", api.handlePutBucketPublicExposure)
		})

		r.Route("/buckets/{bucket}/governance/protection", func(r chi.Router) {
			r.Get("/", api.handleGetBucketProtection)
			r.Put("/", api.handlePutBucketProtection)
		})

		r.Route("/buckets/{bucket}/governance/versioning", func(r chi.Router) {
			r.Get("/", api.handleGetBucketVersioning)
			r.Put("/", api.handlePutBucketVersioning)
		})

		r.Route("/buckets/{bucket}/governance/encryption", func(r chi.Router) {
			r.Get("/", api.handleGetBucketEncryption)
			r.Put("/", api.handlePutBucketEncryption)
		})

		r.Route("/buckets/{bucket}/governance/lifecycle", func(r chi.Router) {
			r.Get("/", api.handleGetBucketLifecycle)
			r.Put("/", api.handlePutBucketLifecycle)
		})

		r.Route("/buckets/{bucket}/governance/sharing", func(r chi.Router) {
			r.Get("/", api.handleGetBucketSharing)
			r.Put("/", api.handlePutBucketSharing)
		})

		r.Route("/buckets/{bucket}/objects", func(r chi.Router) {
			r.Get("/", api.handleListObjects)
			r.Delete("/", api.handleDeleteObjects)
		})
		r.Get("/buckets/{bucket}/objects/meta", api.handleGetObjectMeta)
		r.Post("/buckets/{bucket}/objects/folder", api.handleCreateObjectFolder)
		r.Get("/buckets/{bucket}/objects/download", api.handleDownloadObject)
		r.Get("/buckets/{bucket}/objects/download-url", api.handleGetObjectDownloadURL)
		r.Get("/buckets/{bucket}/objects/favorites", api.handleListObjectFavorites)
		r.Get("/buckets/{bucket}/objects/thumbnail", api.handleGetObjectThumbnail)

		r.Route("/uploads", func(r chi.Router) {
			r.Post("/", api.handleCreateUploadSession)
			r.Post("/{uploadId}/presign", api.handlePresignUpload)
			r.Post("/{uploadId}/multipart/complete", api.handleCompleteMultipartUpload)
			r.Post("/{uploadId}/multipart/abort", api.handleAbortMultipartUpload)
			r.Get("/{uploadId}/chunks", api.handleGetUploadChunks)
			r.Post("/{uploadId}/chunks/batch", api.handleGetUploadChunksBatch)
			r.Post("/{uploadId}/commit", api.handleCommitUpload)
			r.Delete("/{uploadId}", api.handleDeleteUploadSession)
		})

		r.Post("/jobs", api.handleCreateJob)
		r.Delete("/jobs/{jobId}", api.handleDeleteJob)
		r.Post("/jobs/{jobId}/retry", api.handleRetryJob)
		r.Post("/jobs/{jobId}/cancel", api.handleCancelJob)
	})

	r.Mount("/api/v1", apiRouter)

	r.With(api.requireLocalHost).Get("/download-proxy", api.handleDownloadProxy)
	r.With(api.requireLocalHost).Head("/download-proxy", api.handleDownloadProxy)

	r.With(api.requireLocalHost).Get("/openapi.yml", func(w http.ResponseWriter, r *http.Request) {
		specPath, ok := findOpenAPISpecPath(dep.Config.StaticDir)
		if !ok {
			http.NotFound(w, r)
			return
		}
		w.Header().Set("Content-Type", "application/yaml; charset=utf-8")
		http.ServeFile(w, r, specPath)
	})
	r.With(api.requireLocalHost).Get("/docs", serveOpenAPIDocs)
	r.With(api.requireLocalHost).Get("/docs/", serveOpenAPIDocs)

	r.With(api.requireLocalPeer).Get("/healthz", api.handleHealthz)
	r.With(api.requireLocalPeer).Get("/readyz", api.handleReadyz)
	r.With(api.requireLocalPeer).Get("/workerz", api.handleWorkerz)
	r.With(api.requireLocalPeer, api.requireAPIToken).Get("/metrics", api.handleMetrics)

	staticIndex := filepath.Join(dep.Config.StaticDir, "index.html")
	uiEnabled := false
	if dep.Config.StaticDir != "" {
		if info, err := os.Stat(staticIndex); err == nil && !info.IsDir() {
			spa := spaHandler(dep.Config.StaticDir)
			r.With(api.requireLocalHost).Get("/", spa)
			r.With(api.requireLocalHost).Get("/*", spa)
			r.With(api.requireLocalHost).Head("/", spa)
			r.With(api.requireLocalHost).Head("/*", spa)
			uiEnabled = true
		}
	}

	if !uiEnabled {
		r.With(api.requireLocalHost).Get("/", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "text/plain; charset=utf-8")
			_, _ = w.Write([]byte("s3desk backend is running\n\nHint: build the frontend and point --static-dir to ../frontend/dist\n"))
		})
	}

	return r
}

func spaHandler(staticDir string) http.HandlerFunc {
	fileServer := http.FileServer(http.Dir(staticDir))
	indexPath := filepath.Join(staticDir, "index.html")

	return func(w http.ResponseWriter, r *http.Request) {
		if r.Method != http.MethodGet && r.Method != http.MethodHead {
			http.NotFound(w, r)
			return
		}

		reqPath := path.Clean(r.URL.Path)
		if reqPath == "/" {
			w.Header().Set("Cache-Control", "no-cache")
			http.ServeFile(w, r, indexPath)
			return
		}

		rel := strings.TrimPrefix(reqPath, "/")
		target := filepath.Join(staticDir, rel)
		if info, err := os.Stat(target); err == nil && !info.IsDir() {
			if strings.HasPrefix(rel, "assets/") {
				w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
			} else {
				w.Header().Set("Cache-Control", "public, max-age=86400")
			}
			fileServer.ServeHTTP(w, r)
			return
		}

		if rel == "assets" || strings.HasPrefix(rel, "assets/") {
			w.Header().Set("Cache-Control", "no-cache")
			http.NotFound(w, r)
			return
		}

		w.Header().Set("Cache-Control", "no-cache")
		http.ServeFile(w, r, indexPath)
	}
}

func findOpenAPISpecPath(staticDir string) (path string, ok bool) {
	candidates := []string{}
	if exe, err := os.Executable(); err == nil {
		exeDir := filepath.Dir(exe)
		candidates = append(candidates, filepath.Join(exeDir, "openapi.yml"))
	}

	candidates = append(candidates,
		"openapi.yml",
		filepath.Join("dist", "openapi.yml"),
		filepath.Join("..", "openapi.yml"),
	)

	if staticDir != "" {
		sd := filepath.Clean(staticDir)
		candidates = append(candidates,
			filepath.Join(sd, "..", "openapi.yml"),
			filepath.Join(sd, "..", "..", "openapi.yml"),
		)
	}

	for _, p := range candidates {
		if fileExists(p) {
			return p, true
		}
	}
	return "", false
}

func fileExists(path string) bool {
	info, err := os.Stat(path)
	return err == nil && !info.IsDir()
}