# openapi.yml
openapi: 3.0.3
info:
  title: S3Desk API
  version: 0.1.0
  description: |
    Draft API spec for S3Desk, a local-only (127.0.0.1) dashboard targeting S3-compatible object storage.
    - Metadata/browsing and downloads are powered by `rclone`.
    - Bulk transfers run as Jobs powered by `rclone`.
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Default local server
...
paths:
  /ws:
    get:
      tags: [Events]
      summary: WebSocket event stream
      description: |
        Upgrades to a WebSocket and streams `job.*` events.

        Browsers should first request a short-lived realtime ticket via `POST /realtime-ticket`
        and then connect with `?realtimeTicket=...`.

        Optional: `afterSeq` can be used to replay buffered (non-log) events after a given sequence number.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: realtimeTicket
          in: query
          required: false
          schema: { type: string }
        - name: afterSeq
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 0 }
        - name: includeLogs
          in: query
          required: false
          description: Whether to include `job.log` events in the stream.
          schema: { type: boolean, default: true }
      responses:
        "101":
          description: Switching Protocols
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
  /events:
    get:
      tags: [Events]
      summary: Server-sent events (SSE) stream
      description: |
        Streams `job.*` events using `text/event-stream` (one JSON event per `data:` line).

        Browsers should first request a short-lived realtime ticket via `POST /realtime-ticket`
        and then connect with `?realtimeTicket=...`.

        Each message includes an SSE `id:` equal to the server-side event `seq`. When reconnecting, clients may send `Last-Event-ID` (or `afterSeq`) to receive buffered (non-log) events that were missed.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: realtimeTicket
          in: query
          required: false
          schema: { type: string }
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: integer, format: int64, minimum: 0 }
        - name: afterSeq
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 0 }
        - name: includeLogs
          in: query
          required: false
          description: Whether to include `job.log` events in the stream.
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /realtime-ticket:
    post:
      tags: [Events]
      summary: Issue a short-lived realtime ticket
      description: |
        Issues a single-use realtime ticket for browser WebSocket or SSE connections.
        Clients should call this authenticated endpoint first, then connect to `/ws`
        or `/events` with `?realtimeTicket=...`.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: transport
          in: query
          required: true
          schema:
            type: string
            enum: [ws, sse]
      responses:
        "200":
          description: OK (blocked by preflight; no data written)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerPortableImportResponse"
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [ticket, transport, expiresAt, expiresInSeconds]
                properties:
                  ticket:
                    type: string
                  transport:
                    type: string
                    enum: [ws, sse]
                  expiresAt:
                    type: string
                    format: date-time
                  expiresInSeconds:
                    type: integer
                    format: int64
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /bootstrap:
    get:
      tags: [System]
      summary: Authenticated UI bootstrap data
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /meta:
    get:
      tags: [System]
      summary: Server metadata
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
  /server/backup:
    get:
      tags: [System]
      summary: Download a server backup bundle
      description: |
        Exports a `.tar.gz` server backup bundle containing the sqlite snapshot and a
        selected slice of `DATA_DIR` runtime state. Use `scope=full` for a full local
        backup, or `scope=cache_metadata` for a lighter bundle containing metadata and
        cache state such as thumbnails. Use `scope=portable` to export a logical
        database-neutral migration bundle for portable import flows. Use
        `confidentiality=encrypted` to encrypt the payload with
        a non-empty `X-S3Desk-Backup-Password` when supplied, otherwise with the
        current `ENCRYPTION_KEY`, while keeping the outer bundle metadata clear.
        Restore/import endpoints do not read the export header; supply password-protected
        bundles through their multipart `password` field or matching UI input. During
        restore/import, a non-empty multipart password is used before destination
        server `ENCRYPTION_KEY` fallback. Snapshot-style export currently supports
        sqlite-backed servers only, while portable export supports sqlite and postgres
        source deployments.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: X-S3Desk-Backup-Password
          in: header
          required: false
          description: Optional operator-supplied export password used when `confidentiality=encrypted`. Restore/import endpoints do not read this header; provide the same value in their multipart `password` field when processing password-protected bundles.
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
            enum: [full, cache_metadata, portable]
            default: full
          description: Selects whether the bundle contains a sqlite snapshot export (`full`, `cache_metadata`) or a logical portable migration bundle (`portable`).
        - name: includeThumbnails
          in: query
          schema:
            type: boolean
            default: true
          description: When `scope=portable`, includes thumbnail cache assets in the bundle.
        - name: confidentiality
          in: query
          schema:
            type: string
            enum: [clear, encrypted]
            default: clear
          description: Encrypts the bundle payload with a non-empty `X-S3Desk-Backup-Password` when supplied, otherwise with the current `ENCRYPTION_KEY`, while leaving the outer tar.gz manifest readable.
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              description: Suggested filename for the backup bundle.
              schema: { type: string }
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /server/restore:
    post:
      tags: [System]
      summary: Stage a server backup bundle for restore
      description: |
        Uploads a `.tar.gz` bundle previously created by `/server/backup` and stages it
        under `DATA_DIR/restores/<id>` without overwriting the live instance. The staged
        directory can then be used as the destination server's `DATA_DIR`. The response
        includes a stage-only summary plus an apply plan and helper command for cutover.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [bundle]
              properties:
                bundle:
                  type: string
                  format: binary
                password:
                  type: string
                  description: Optional multipart password field used when the backup payload was exported with operator-supplied password protection; omit it only when restoring an encrypted bundle with the destination server encryption key.
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRestoreResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
  /server/backup/transfer:
    post:
      tags: [System]
      summary: Export a server backup to remote storage
      description: Creates the same validated backup bundle as `GET /server/backup`, then copies it to an existing object-storage profile, an FTP server, or an NFS path already mounted on the server. NFS paths are constrained by `ALLOWED_LOCAL_DIRS`. FTP is plain FTP; use encrypted backup confidentiality when the network is not trusted.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerBackupTransferRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerBackupTransferResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /server/restore/transfer:
    post:
      tags: [System]
      summary: Fetch and stage a server backup from remote storage
      description: Fetches one backup bundle from an existing object-storage profile, FTP server, or mounted NFS path, enforces the restore size limit, and passes it through the same checksum, signature, encryption, preflight, and stage-only restore flow as `POST /server/restore`.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerRestoreTransferRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRestoreResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /server/import-portable/preview:
    post:
      tags: [System]
      summary: Preview a portable backup import
      description: |
        Uploads a portable `.tar.gz` bundle and performs a dry-run compatibility check
        without writing database state.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [bundle]
              properties:
                bundle:
                  type: string
                  format: binary
                password:
                  type: string
                  description: Optional multipart password field used when the portable payload was exported with operator-supplied password protection; omit it only when importing an encrypted bundle with the destination server encryption key.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerPortableImportResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
  /server/import-portable:
    post:
      tags: [System]
      summary: Import a portable backup bundle
      description: |
        Uploads a portable `.tar.gz` bundle and imports its logical application data into
        the current database backend using replace semantics.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [bundle]
              properties:
                bundle:
                  type: string
                  format: binary
                password:
                  type: string
                  description: Optional multipart password field used when the portable payload was exported with operator-supplied password protection; omit it only when importing an encrypted bundle with the destination server encryption key.
      responses:
        "200":
          description: OK (blocked by preflight; no data written)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerPortableImportResponse"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerPortableImportResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
  /server/restores:
    get:
      tags: [System]
      summary: List staged restore bundles
      description: |
        Lists staged restore directories under `DATA_DIR/restores`. These are safe-to-delete
        staged bundles created by `/server/restore`; deleting them does not change the live
        `DATA_DIR`.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStagedRestoreListResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /server/restores/{restoreId}:
    delete:
      tags: [System]
      summary: Delete a staged restore bundle
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: restoreId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /profiles:
    get:
      tags: [Profiles]
      summary: List profiles
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      tags: [Profiles]
      summary: Create profile
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /profiles/{profileId}:
    patch:
      tags: [Profiles]
      summary: Update profile
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - $ref: "#/components/parameters/ProfileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Profiles]
      summary: Delete profile
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
...
  /buckets:
    get:
      tags: [Buckets]
      summary: List buckets
      parameters:
        - $ref: "#/components/parameters/XProfileId"
        - $ref: "#/components/parameters/XApiToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bucket"
        "400":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Buckets]
      summary: Create bucket
      parameters:
        - $ref: "#/components/parameters/XProfileId"
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BucketCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bucket"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /buckets/{bucket}:
    delete:
      tags: [Buckets]
      summary: Delete bucket
      description: |
        Deletes an empty bucket. If the bucket is not empty, the server returns `409 bucket_not_empty`.
      parameters:
        - $ref: "#/components/parameters/XProfileId"
        - $ref: "#/components/parameters/XApiToken"
        - $ref: "#/components/parameters/BucketName"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
...
  /buckets/{bucket}/policy/validate:
    post:
      tags: [Buckets]
      summary: Validate bucket policy
      description: |
        Performs static (non-mutating) validation of a provider-specific bucket access policy document.

        - For S3 providers, this is syntactic lint only. Real provider validation happens on PUT.
        - For GCS/Azure providers, this validates the expected JSON shape for IAM/ACL documents.
      parameters:
        - $ref: "#/components/parameters/XProfileId"
        - $ref: "#/components/parameters/XApiToken"
        - $ref: "#/components/parameters/BucketName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BucketPolicyPutRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketPolicyValidateResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /buckets/{bucket}/governance:
    get:
      tags: [Buckets]
      summary: Get normalized bucket governance summary
      description: |
        Returns the provider-aware bucket governance summary used by the typed controls UI.

        The initial rollout is AWS-first and returns a normalized capability map plus advanced policy affordances.
      parameters:
        - $ref: "#/components/parameters/XProfileId"
        - $ref: "#/components/parameters/XApiToken"
        - $ref: "#/components/parameters/BucketName"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketGovernanceView"
        "400":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
...
    XApiToken:
      name: X-Api-Token
      in: header
      required: false
      description: Optional local API token to mitigate localhost/CSRF style attacks.
      schema: { type: string }
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    BootstrapResponse:
      type: object
      required: [meta, profiles]
      properties:
        meta: { $ref: "#/components/schemas/MetaResponse" }
        profiles:
          type: array
          items: { $ref: "#/components/schemas/Profile" }
...
    ServerStagedRestoreListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServerStagedRestore"
...
    ServerPortableImportResponse:
      type: object
      required: [manifest, mode, targetDbBackend, preflight, entities, verification]
      properties:
        manifest:
          $ref: "#/components/schemas/ServerMigrationManifest"
        mode: { type: string }
        targetDbBackend: { type: string }
        preflight:
          $ref: "#/components/schemas/ServerPortableImportPreflight"
        entities:
          type: array
          items:
            $ref: "#/components/schemas/ServerPortableImportEntityResult"
        verification:
          $ref: "#/components/schemas/ServerPortableImportVerification"
        assetStagingDir: { type: string }
        warnings:
          type: array
          items: { type: string }
...